-Edit and improved the Data base Schema by adding athleteprofile, companyprofile, and college table -Created us_univeristes.cvs to import colleges to college table -modify and added more functionality to admin dashboard

+++ app.py
@@ -1,12 +1,15 @@
 from flask import Flask, render_template, request, redirect, url_for, flash
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import LoginManager, UserMixin, login_user, login_required, logout_user, current_user
-from sqlalchemy.dialects.postgresql import ENUM
 from werkzeug.utils import secure_filename
 from sqlalchemy import and_
 from werkzeug.security import check_password_hash
 from PIL import Image
-from sqlalchemy.orm import relationship
+from sqlalchemy import Column, Integer, String, Text, ForeignKey, Boolean, Enum, TIMESTAMP, create_engine
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import relationship, backref
+from flask import jsonify
+from sqlalchemy.orm import joinedload
 import os
 
 app = Flask(__name__)
@@ -20,7 +23,7 @@ login_manager = LoginManager()
 login_manager.login_view = 'login'
 login_manager.init_app(app)
 
-UPLOAD_FOLDER = 'static/profile_picture'
+UPLOAD_FOLDER = os.path.join('static', 'profile_picture')
 ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif'}
 
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
@@ -30,9 +33,9 @@ class User(UserMixin, db.Model):
     userid = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(50), unique=True, nullable=False)
     password = db.Column(db.String(50), nullable=False)
-    usertype = db.Column(ENUM('Admin', 'Company', 'Athlete', name='usertype'), nullable=False)
+    usertype = db.Column(Enum('Admin', 'Company', 'Athlete', name='usertype'), nullable=False)
     email = db.Column(db.String(100), unique=True, nullable=False)
-    profile = relationship('Profile', backref='appuser', lazy=True, uselist=False)
+    profile = relationship('Profile', backref='appuser', uselist=False)
 
     def get_id(self):
         return str(self.userid)
@@ -40,38 +43,58 @@ class User(UserMixin, db.Model):
 class Profile(db.Model):
     __tablename__ = 'profile'
     profileid = db.Column(db.Integer, primary_key=True)
-    userid = db.Column(db.Integer, db.ForeignKey('appuser.userid'), unique=True)
+    userid = db.Column(db.Integer, ForeignKey('appuser.userid'), unique=True)
     fullname = db.Column(db.String(100))
     bio = db.Column(db.Text)
-    gender = db.Column(ENUM('Male', 'Female', 'Other', name='gendertype'))
-    sportscategory = db.Column(db.String(50))
     profilepicture = db.Column(db.Text)
     verifiedstatus = db.Column(db.Boolean, default=False)
-    userid = db.Column(db.Integer, db.ForeignKey('appuser.userid'), unique=True)
+    athlete_profile = relationship('AthleteProfile', backref='profile', uselist=False)
+
+
+class College(db.Model):
+    __tablename__ = 'college'
+    collegeid = db.Column(db.Integer, primary_key=True)
+    collegename = db.Column(db.String(100), unique=True, nullable=False)
+    athletes = relationship('AthleteProfile', backref='college')
+
+class AthleteProfile(db.Model):
+    __tablename__ = 'athleteprofile'
+    athleteprofileid = db.Column(db.Integer, primary_key=True)
+    profileid = db.Column(db.Integer, db.ForeignKey('profile.profileid', ondelete='CASCADE'), unique=True)
+    gender = db.Column(Enum('Male', 'Female', 'Other', name='gendertype'))
+    sportscategory = db.Column(Enum('Basketball', 'Football', 'Soccer', name='sportscategorytype'))
+    collegeid = db.Column(db.Integer, ForeignKey('college.collegeid'))
+
+
+class CompanyProfile(db.Model):
+    __tablename__ = 'companyprofile'
+    companyprofileid = db.Column(db.Integer, primary_key=True)
+    profileid = db.Column(db.Integer, ForeignKey('profile.profileid'), unique=True)
+    companyname = db.Column(db.String(100), nullable=False)
+    companylogo = db.Column(db.Text)
 
 class Message(db.Model):
     __tablename__ = 'message'
-    MessageID = db.Column(db.Integer, primary_key=True)
-    SenderID = db.Column(db.Integer, db.ForeignKey('appuser.UserID'))
-    ReceiverID = db.Column(db.Integer, db.ForeignKey('appuser.UserID'))
-    Content = db.Column(db.Text)
-    Timestamp = db.Column(db.DateTime, default=db.func.current_timestamp())
+    messageid = db.Column(db.Integer, primary_key=True)
+    senderid = db.Column(db.Integer, ForeignKey('appuser.userid'))
+    receiverid = db.Column(db.Integer, ForeignKey('appuser.userid'))
+    content = db.Column(db.Text)
+    timestamp = db.Column(db.TIMESTAMP, default=db.func.current_timestamp())
 
 class Offer(db.Model):
     __tablename__ = 'offer'
-    OfferID = db.Column(db.Integer, primary_key=True)
-    CompanyID = db.Column(db.Integer, db.ForeignKey('profile.ProfileID'))
-    AthleteID = db.Column(db.Integer, db.ForeignKey('profile.ProfileID'))
-    Details = db.Column(db.Text)
-    Status = db.Column(ENUM('Pending', 'Accepted', 'Declined', 'Counter-offered', name='offerstatus'), default='Pending')
+    offerid = db.Column(db.Integer, primary_key=True)
+    companyid = db.Column(db.Integer, ForeignKey('companyprofile.companyprofileid'))
+    athleteid = db.Column(db.Integer, ForeignKey('athleteprofile.athleteprofileid'))
+    details = db.Column(db.Text)
+    status = db.Column(Enum('Pending', 'Accepted', 'Declined', 'Counter-offered', name='offerstatus'), default='Pending')
 
 class Watchlist(db.Model):
     __tablename__ = 'watchlist'
-    WatchlistID = db.Column(db.Integer, primary_key=True)
-    CompanyID = db.Column(db.Integer, db.ForeignKey('profile.ProfileID'))
-    AthleteID = db.Column(db.Integer, db.ForeignKey('profile.ProfileID'))
-    SportsCategory = db.Column(db.String(50))
-
+    watchlistid = db.Column(db.Integer, primary_key=True)
+    companyid = db.Column(db.Integer, ForeignKey('companyprofile.companyprofileid'))
+    athleteid = db.Column(db.Integer, ForeignKey('athleteprofile.athleteprofileid'))
+    sportscategory = db.Column(Enum('Basketball', 'Football', 'Soccer', name='sportscategorytype'))
 @app.route('/', methods=['GET'])
 def index():
     return render_template('index.html')
@@ -82,24 +105,44 @@ def load_user(user_id):
 @app.route('/login', methods=['GET', 'POST'])
 def login():
     if current_user.is_authenticated:
-        return redirect(url_for('dashboard'))
+        if current_user.usertype == 'Admin':
+            return redirect(url_for('admin_dashboard'))
+        else:
+            return redirect(url_for('dashboard'))
+
     if request.method == 'POST':
         username = request.form.get('username')
         password = request.form.get('password')
         user = User.query.filter_by(username=username).first()
         if user and user.password == password:
             login_user(user)
-            return redirect(url_for('dashboard'))
+            if user.usertype == 'Admin':
+                return redirect(url_for('admin_dashboard'))
+            else:
+                return redirect(url_for('dashboard'))
         else:
             flash('Login Unsuccessful. Please check your username and password', 'danger')
     return render_template('login.html')
 
+
 @app.route('/dashboard')
 @login_required
 def dashboard():
-    profile_exists = Profile.query.filter_by(userid=current_user.userid).first() is not None
+    if current_user.usertype == 'Admin':
+        return redirect(url_for('admin_dashboard'))
+
     profile = Profile.query.filter_by(userid=current_user.userid).first()
-    return render_template('dashboard.html', profile_exists=profile_exists, profile=profile)
+    profile_exists = profile is not None
+    athlete_profile = None
+    company_profile = None
+
+    if profile_exists:
+        if current_user.usertype == 'Athlete':
+            athlete_profile = AthleteProfile.query.filter_by(profileid=profile.profileid).first()
+        elif current_user.usertype == 'Company':
+            company_profile = CompanyProfile.query.filter_by(profileid=profile.profileid).first()
+
+    return render_template('dashboard.html', profile=profile, athlete_profile=athlete_profile, company_profile=company_profile, profile_exists=profile_exists)
 
 @app.route('/logout', methods=['GET'])
 @login_required
@@ -147,40 +190,47 @@ def create_profile():
     if existing_profile:
         flash('You already have a profile!', 'info')
         return redirect(url_for('dashboard'))
+
     if request.method == 'POST':
         full_name = request.form.get('full_name')
         bio = request.form.get('bio')
-        gender = request.form.get('gender')
-        sports_category = request.form.get('sports_category')
 
-        profile_picture = request.files.get('profile_picture')
-        filepath = None
-        if profile_picture and allowed_file(profile_picture.filename):
-            filename = secure_filename(profile_picture.filename)
-            filepath = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-
-            with Image.open(profile_picture) as img:
-                img = img.resize((256, 256))
-                img.save(filepath)
-
-            # Updated code to save relative path into the database
-            relative_filepath = os.path.join('profile_picture', filename)
-            filepath = relative_filepath  # overwrite the filepath variable with the relative path
+        college_id = request.form.get('college')
+        college = College.query.get(college_id)
+        if not college:
+            flash('Selected college does not exist!', 'danger')
+            return redirect(url_for('create_profile'))
 
         new_profile = Profile(
             userid=current_user.userid,
             fullname=full_name,
-            bio=bio,
-            gender=gender,
-            sportscategory=sports_category,
-            profilepicture=filepath  # this now uses the relative path
+            bio=bio
         )
         db.session.add(new_profile)
+        db.session.flush()
+
+        athlete_profile = AthleteProfile(
+            profileid=new_profile.profileid,
+            gender=request.form.get('gender'),
+            sportscategory=request.form.get('sports_category'),
+            collegeid=college_id
+        )
+        db.session.add(athlete_profile)
         db.session.commit()
+
         flash('Profile created successfully!', 'success')
         return redirect(url_for('dashboard'))
 
-    return render_template('create_profile.html')
+    return render_template('create_profile.html', user_type=current_user.usertype)
+
+
+@app.route('/get_colleges')
+def get_colleges():
+    search_term = request.args.get('q', '')
+    matching_colleges = College.query.filter(College.collegename.ilike(f'%{search_term}%')).all()
+    college_list = [{'id': college.collegeid, 'text': college.collegename} for college in matching_colleges]
+
+    return jsonify(results=college_list)
 
 
 @app.route('/edit_profile', methods=['GET', 'POST'])
@@ -191,32 +241,36 @@ def edit_profile():
         flash('No profile found!', 'danger')
         return redirect(url_for('dashboard'))
 
+    athlete_profile = AthleteProfile.query.filter_by(profileid=profile.profileid).first()
+
     if request.method == 'POST':
         profile.fullname = request.form.get('full_name')
         profile.bio = request.form.get('bio')
-        profile.gender = request.form.get('gender')
-        profile.sportscategory = request.form.get('sports_category')
 
         profile_picture = request.files.get('profile_picture')
         if profile_picture and allowed_file(profile_picture.filename):
             filename = secure_filename(profile_picture.filename)
             filepath = os.path.join(app.config['UPLOAD_FOLDER'], filename)
 
-
             with Image.open(profile_picture) as img:
-
                 img = img.resize((256, 256))
-
                 img.save(filepath)
 
             relative_filepath = os.path.join('profile_picture', filename)
             profile.profilepicture = relative_filepath
 
+        if current_user.usertype == 'Athlete':
+            athlete_profile.gender = request.form.get('gender')
+            athlete_profile.sportscategory = request.form.get('sports_category')
+            athlete_profile.collegeid = request.form.get('college')
+
         db.session.commit()
         flash('Profile updated successfully!', 'success')
         return redirect(url_for('dashboard'))
 
-    return render_template('edit_profile.html', profile=profile)
+    colleges = College.query.all()
+    return render_template('edit_profile.html', profile=profile, athlete_profile=athlete_profile, colleges=colleges)
+
 
 def allowed_file(filename):
     return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
@@ -243,20 +297,28 @@ def upload_file():
 @app.route('/delete_account', methods=['GET', 'POST'])
 @login_required
 def delete_account():
-    profile = Profile.query.filter_by(userid=current_user.userid).first()
     user = User.query.get(current_user.userid)
 
     if request.method == 'POST':
+        if user:
+            profile = Profile.query.filter_by(userid=current_user.userid).first()
 
-        if profile:
-            db.session.delete(profile)
+            if profile:
+                athlete_profile = AthleteProfile.query.filter_by(profileid=profile.profileid).first()
+                company_profile = CompanyProfile.query.filter_by(profileid=profile.profileid).first()
 
+                if athlete_profile:
+                    db.session.delete(athlete_profile)
+                elif company_profile:
+                    db.session.delete(company_profile)
 
-        db.session.delete(user)
-        db.session.commit()
+                db.session.delete(profile)
 
-        flash('Your account and profile have been deleted', 'success')
-        return redirect(url_for('goodbye'))
+            db.session.delete(user)
+            db.session.commit()
+
+            flash('Your account and profile have been deleted', 'success')
+            return redirect(url_for('goodbye'))
 
     return render_template('delete_account.html')
 @app.route('/goodbye')
@@ -287,27 +349,32 @@ def edit_user(user_id):
         flash('User not found', 'danger')
         return redirect(url_for('admin_dashboard'))
 
+    colleges = College.query.all()
+
     if request.method == 'POST':
         user.username = request.form.get('username')
         user.email = request.form.get('email')
+
         new_password = request.form.get('password')
-
-
         if new_password:
             user.password = new_password
 
 
-        new_usertype = request.form.get('usertype')
-        if new_usertype and new_usertype in ['Admin', 'Company', 'Athlete']:
-            user.usertype = new_usertype
+        if user.usertype == 'Athlete' and user.profile and user.profile.athlete_profile:
+            gender = request.form.get('gender')
+            sportscategory = request.form.get('sportscategory')
+            collegeid = request.form.get('collegeid')
 
+            athlete_profile = user.profile.athlete_profile
+            athlete_profile.gender = gender
+            athlete_profile.sportscategory = sportscategory
+            athlete_profile.collegeid = collegeid
 
         db.session.commit()
         flash(f'Successfully edited {user.username}', 'success')
         return redirect(url_for('admin_dashboard'))
 
-
-    return render_template('user_edit.html', user=user)
+    return render_template('user_edit.html', user=user, colleges=colleges)
 
 @app.route('/admin/delete_user/<int:user_id>')
 @login_required
@@ -359,9 +426,65 @@ def user_list_page():
         return redirect(url_for('index'))
 
     search = request.form.get('search')
-    usertype_filter = request.form.get('usertype_filter')
     gender_filter = request.form.get('gender_filter')
     verification_filter = request.form.get('verification_filter')
+
+    filter_conditions = [User.usertype == 'Athlete']
+
+    if search:
+        filter_conditions.append(User.username.ilike(f"%{search}%"))
+
+    if gender_filter:
+        filter_conditions.append(AthleteProfile.gender == gender_filter)
+
+    if verification_filter:
+        verified_status = True if verification_filter == "Verified" else False
+        filter_conditions.append(Profile.verifiedstatus == verified_status)
+
+    users = User.query\
+                .join(Profile, User.userid == Profile.userid, isouter=True)\
+                .join(AthleteProfile, Profile.profileid == AthleteProfile.profileid, isouter=True)\
+                .filter(and_(*filter_conditions)).all()
+
+    return render_template('user_list.html', users=users)
+
+@app.route('/admin/company_user_list', methods=['GET', 'POST'])
+@login_required
+def company_user_list_page():
+    if current_user.usertype != 'Admin':
+        flash('Permission Denied: You are not an administrator', 'danger')
+        return redirect(url_for('index'))
+
+    search = request.form.get('search')
+    verification_filter = request.form.get('verification_filter')
+
+    filter_conditions = []
+
+    if search:
+        filter_conditions.append(CompanyProfile.companyname.ilike(f"%{search}%"))
+
+    if verification_filter:
+        verified_status = True if verification_filter == "Verified" else False
+        filter_conditions.append(Profile.verifiedstatus == verified_status)
+
+
+    company_users = User.query.filter(User.usertype == 'Company')
+
+    users = company_users\
+                .join(Profile, User.userid == Profile.userid, isouter=True)\
+                .join(CompanyProfile, Profile.profileid == CompanyProfile.profileid, isouter=True)\
+                .filter(*filter_conditions).all()
+
+    return render_template('company_user_list.html', users=users)
+
+@app.route('/admin/admin_user_list', methods=['GET', 'POST'])
+@login_required
+def admin_user_list_page():
+    if current_user.usertype != 'Admin':
+        flash('Permission Denied: You are not an administrator', 'danger')
+        return redirect(url_for('index'))
+
+    search = request.form.get('search')
 
 
     filter_conditions = []
@@ -369,23 +492,16 @@ def user_list_page():
     if search:
         filter_conditions.append(User.username.ilike(f"%{search}%"))
 
-    if usertype_filter:
-        filter_conditions.append(User.usertype == usertype_filter)
-
-    if gender_filter:
-
-        filter_conditions.append(Profile.gender == gender_filter)
-
-    if verification_filter:
-        verified_status = True if verification_filter == "Verified" else False
-        filter_conditions.append(Profile.verifiedstatus == verified_status)
 
 
 
-    users = User.query.join(Profile, User.userid == Profile.userid, isouter=True).filter(and_(*filter_conditions)).all()
+    admin_users = User.query.filter(User.usertype == 'Admin')
 
-    return render_template('user_list.html', users=users)
+    users = admin_users\
+                .join(Profile, User.userid == Profile.userid, isouter=True)\
+                .filter(*filter_conditions).all()
 
+    return render_template('admin_user_list.html', users=users)
 
 @app.route('/admin/add_admin', methods=['GET', 'POST'])
 @login_required
@@ -407,6 +523,34 @@ def add_admin():
         return redirect(url_for('admin_dashboard'))
 
     return render_template('add_admin.html')
+
+
+@app.route('/admin_edit/<int:user_id>', methods=['GET', 'POST'])
+@login_required
+def admin_edit(user_id):
+    user = User.query.get(user_id)
+    if not user:
+        flash('User not found', 'danger')
+        return redirect(url_for('admin_dashboard'))
+
+    if request.method == 'POST':
+
+        user.username = request.form['username']
+        user.password = request.form['password']
+        user.email = request.form['email']
+
+        try:
+            db.session.commit()
+            flash('Profile successfully updated', 'success')
+        except Exception as e:
+            db.session.rollback()
+            flash(f'Error updating profile: {str(e)}', 'danger')
+
+        return redirect(url_for('admin_dashboard'))
+
+    return render_template('admin_edit.html', user=user)
+
+
 
 @app.route('/aboutus')
 def about_us():
